gemini/pyproject.toml
[project]
name = "simple-gemini-agent"
version = "0.1.0"
description = "A simple interactive AI agent using Gemini with tool calling"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]
